dl.R
swapped interactive with ask (which was not used) which is v important to be able to leave code running without being prompted.

Changed default file to second. The only year where two csvs are available is 2020 and for escooter data. Maybe this is interesting for some, but the other file should be the default.

returning destfile works better for dependent functions

get.R

main change is to skip the read_ functions which involve multiple nested functions appear to be unneccesary and are a headache to debug.

utils.R

Main change is to fix find_file_name function which was looking for a csv with year name for years before 2018. This data is all in one file which has 1979 in.